+++ OneSignalExample/Assets/OneSignal/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `InstallEdm4uStep` now imports version [1.2.169](https://github.com/googlesamples/unity-jar-resolver/releases/tag/v1.2.169) of [EDM4U](https://github.com/googlesamples/unity-jar-resolver)
 - Log an error in the example app when `RequiresPrivacyConsent` is attempted to be set to false from true
 - Internal state mappings on iOS now rely on class defined objects over dynamic Dictionary types
+- Replaced manual manipulation of iOS entitlements in post processing with Unity's [ProjectCapabilityManager](https://docs.unity3d.com/ScriptReference/iOS.Xcode.ProjectCapabilityManager.html)
 
 ## [3.0.0-beta.5]
 ### Changed